move FileDecorator one directory up, since always only 1 class
